bulk.c (tds_bcp_send_record): Reject invalid NULL data.
When the user erroneously supplies NULL data for a column which does not
use a nullable type and is not otherwise known to be nullable, cleanly
fail.  Pressing on anyway (in hopes that a server-side default would
take effect?) was liable to trigger legitimate assertion failures in
tds_checks.c.  JIRA: CXX-4155.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@70352 112efe8e-fc92-43c6-89b6-d25c299ce97b